Defines for calibration, battery level, SD logging.
Added defines at the top of main for calibration for diff
bikes. Comment out the blocks unused.

Estimate and update the battery level every 5 min. Not sure
if this really works well or not.

And add SD logging ability, if hardware is present. Can
disable based on define as well.